[Validator] Moved Optional and Required constraints to dedicated sub namespace.
